Remove Oap Data File relevant code
