--- oap-cache/oap/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,12 +1,0 @@
-org.apache.spark.sql.execution.datasources.oap.OapFileFormat
-org.apache.spark.sql.execution.datasources.v2.csv.CSVDataSourceV2
-org.apache.spark.sql.execution.datasources.jdbc.JdbcRelationProvider
-org.apache.spark.sql.execution.datasources.v2.json.JsonDataSourceV2
-org.apache.spark.sql.execution.datasources.noop.NoopDataSource
-org.apache.spark.sql.execution.datasources.orc.OrcFileFormat
-org.apache.spark.sql.execution.datasources.v2.parquet.ParquetDataSourceV2
-org.apache.spark.sql.execution.datasources.v2.text.TextDataSourceV2
-org.apache.spark.sql.execution.streaming.ConsoleSinkProvider
-org.apache.spark.sql.execution.streaming.sources.RateStreamProvider
-org.apache.spark.sql.execution.streaming.sources.TextSocketSourceProvider
-org.apache.spark.sql.execution.datasources.binaryfile.BinaryFileFormat
