+++ oap-cache/oap/src/main/scala/org/apache/spark/sql/OapExtensions.scala
@@ -19,15 +19,11 @@ package org.apache.spark.sql
 
 import org.apache.spark.sql.execution.OapSparkSqlParser
 import org.apache.spark.sql.execution.datasources.OapFileSourceStrategy
-import org.apache.spark.sql.execution.datasources.oap.{OapGroupAggregateStrategy, OapSemiJoinStrategy, OapSortLimitStrategy}
 import org.apache.spark.sql.hive.thriftserver.OapEnv
 
 class OapExtensions extends (SparkSessionExtensions => Unit) {
   override def apply(extensions: SparkSessionExtensions): Unit = {
     // Oap Custom Strategy.
-    extensions.injectPlannerStrategy(_ => OapSortLimitStrategy)
-    extensions.injectPlannerStrategy(_ => OapSemiJoinStrategy)
-    extensions.injectPlannerStrategy(_ => OapGroupAggregateStrategy)
     extensions.injectPlannerStrategy(_ => OapFileSourceStrategy)
     // Oap Custom SqlParser.
     extensions.injectParser((session, delegate) => {
