+++ oap-cache/oap/src/main/scala/org/apache/spark/sql/execution/datasources/oap/utils/CacheStatusSerDe.scala
@@ -22,8 +22,7 @@ import org.json4s.JsonAST._
 import org.json4s.JsonDSL._
 
 import org.apache.spark.sql.execution.datasources.oap.filecache.FiberCacheStatus
-import org.apache.spark.sql.execution.datasources.oap.io.{OapDataFileMeta, OapDataFileMetaV1}
-import org.apache.spark.util.collection.{BitSet, OapBitSet}
+import org.apache.spark.util.collection.OapBitSet
 
 /**
  * This is user defined Json protocol for SerDe, here the format of Json output should like
@@ -69,27 +68,6 @@ private[oap] object CacheStatusSerDe extends SerDe[String, Seq[FiberCacheStatus]
       (word \ "word").extract[Long]
     }.toArray[Long]
     new OapBitSet(words)
-  }
-
-  // we only transfer 4 items in DataFileMeta to driver, ther are rowCountInEachGroup,
-  // rowCountInLastGroup, groupCount, fieldCount respectively
-  private[oap] def dataFileMetaToJson(dataFileMeta: OapDataFileMeta): JValue = {
-    ("rowCountInEachGroup" -> dataFileMeta.rowCountInEachGroup) ~
-      ("rowCountInLastGroup" -> dataFileMeta.rowCountInLastGroup) ~
-      ("groupCount" -> dataFileMeta.groupCount) ~
-      ("fieldCount" -> dataFileMeta.fieldCount)
-  }
-
-  private[oap] def dataFileMetaFromJson(json: JValue): OapDataFileMeta = {
-    val rowCountInEachGroup = (json \ "rowCountInEachGroup").extract[Int]
-    val rowCountInLastGroup = (json \ "rowCountInLastGroup").extract[Int]
-    val groupCount = (json \ "groupCount").extract[Int]
-    val fieldCount = (json \ "fieldCount").extract[Int]
-    new OapDataFileMetaV1(
-      rowCountInEachGroup = rowCountInEachGroup,
-      rowCountInLastGroup = rowCountInLastGroup,
-      groupCount = groupCount,
-      fieldCount = fieldCount)
   }
 
   private[oap] def statusRawDataToJson(statusRawData: FiberCacheStatus): JValue = {
